Move Regex helper files to System.Collections dir
