+++ src/System.Text.RegularExpressions/src/System/Collections/Generic/ValueListBuilder.Pop.cs
