+++ src/System.Text.RegularExpressions/src/System/Collections/HashtableExtensions.cs
@@ -2,9 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System.Collections;
-
-namespace System.Text.RegularExpressions
+namespace System.Collections
 {
     internal static class HashtableExtensions
     {
@@ -15,7 +13,7 @@ namespace System.Text.RegularExpressions
                 value = (T)table[key];
                 return true;
             }
-            value = default(T);
+            value = default;
             return false;
         }
     }
